skills/discord-memo-agent/scripts/requirements.txt
# Discord 메모 에이전트 — 필수 패키지
requests>=2.31.0

# Google Calendar 연동 (선택)
# google-api-python-client>=2.100.0
# google-auth>=2.23.0

python-dotenv>=1.0.0
